Removed transaction statuses and their definitions and referred to specific L2 client implementations
We will revisit the status naming in a separate standard on L2 transaction status nomenclature.

Previous Definitions:
We define the minimal set of transaction statuses as follows:

- Pending: An L2 transaction submitted to an L2, and waiting to be processed
- Replaced: An L2 transaction that was "Pending" was replaced by another L2 transaction
- Dropped: An L2 transaction that was removed from the L2 processing queue
- Success: A Pending L2 transaction has been included into an L2 block
- Finalized: An L2 transaction included in a L2 block that has been finalized on an L1.

Note, that the first four statuses are analogous to L1 transaction statuses. Note also, that there is still the possibility of an L1 reorganization that could revert the status of an L2 transaction back to success. Also at the time of writing, there is no possibility of an L2 reorganization unless through a malicious attacker.

Finally, more fine-grained statuses such as "Pending L1 Finalization" could be introduced to deliver more insights to end-user and improve the user experience.